api_server_django/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
